fix: correct FFI signature for supported_formats and remove icon.png
- Change unhwp_supported_formats return type from pointer to int
- Update Python/C# wrappers to convert bit flags to descriptive string
- Remove PackageIcon reference (icon.png not present)

+++ bindings/csharp/Unhwp/NativeMethods.cs
@@ -133,7 +133,7 @@ namespace Unhwp.Native
         public static extern IntPtr unhwp_version();
 
         [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-        public static extern IntPtr unhwp_supported_formats();
+        public static extern int unhwp_supported_formats();
 
         // Format detection
         [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
